+++ src/main/java/am/aca/wftartproject/service/PurchaseHistoryService.java
@@ -11,6 +11,6 @@ public interface PurchaseHistoryService {
      * @param userId
      * @param itemId
      */
-    void addPurchase(int userId, int itemId);
+    void addPurchase(Long userId, Long itemId);
 
 }
